refactor: rename hook files to have common name both in types and lib
